fixup! refactor(core): move profile_types.ts to primtives
